# frontend/business-warning-system/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Business Warning System API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/action-plan": {
      "get": {
        "tags": [
          "action-plan"
        ],
        "summary": "Get Action Plans",
        "description": "사용자의 모든 개선 계획 조회",
        "operationId": "get_action_plans_api_action_plan_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActionPlan"
                  },
                  "type": "array",
                  "title": "Response Get Action Plans Api Action Plan Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "action-plan"
        ],
        "summary": "Create Action Plan",
        "description": "새로운 개선 계획 생성",
        "operationId": "create_action_plan_api_action_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/action-plan/{plan_id}": {
      "put": {
        "tags": [
          "action-plan"
        ],
        "summary": "Update Action Plan",
        "description": "개선 계획 수정",
        "operationId": "update_action_plan_api_action_plan__plan_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionPlanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "action-plan"
        ],
        "summary": "Delete Action Plan Item",
        "description": "개선 계획 항목 삭제",
        "operationId": "delete_action_plan_item_api_action_plan__plan_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "item_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Get Notifications",
        "description": "알림 목록 조회",
        "operationId": "get_notifications_api_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  },
                  "type": "array",
                  "title": "Response Get Notifications Api Notifications Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/notifications/{notification_id}": {
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "Delete Notification",
        "description": "알림 삭제",
        "operationId": "delete_notification_api_notifications__notification_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/{notification_id}/read": {
      "put": {
        "tags": [
          "notifications"
        ],
        "summary": "Mark Notification As Read",
        "description": "알림을 읽음으로 표시",
        "operationId": "mark_notification_as_read_api_notifications__notification_id__read_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/settings": {
      "put": {
        "tags": [
          "notifications"
        ],
        "summary": "Update Notification Settings",
        "description": "알림 설정 업데이트",
        "operationId": "update_notification_settings_api_notifications_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/user/profile": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get Profile",
        "description": "프로필 조회",
        "operationId": "get_profile_api_user_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update Profile",
        "description": "프로필 수정",
        "operationId": "update_profile_api_user_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__schemas__UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionPlan": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "userId": {
            "type": "integer",
            "title": "Userid"
          },
          "diagnosisId": {
            "type": "string",
            "title": "Diagnosisid"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionPlanItem"
            },
            "type": "array",
            "title": "Items"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "userId",
          "diagnosisId",
          "items",
          "createdAt",
          "updatedAt"
        ],
        "title": "ActionPlan"
      },
      "ActionPlanItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedate"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "priority",
          "status"
        ],
        "title": "ActionPlanItem"
      },
      "ActionPlanRequest": {
        "properties": {
          "diagnosisId": {
            "type": "string",
            "title": "Diagnosisid"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionPlanItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "diagnosisId",
          "items"
        ],
        "title": "ActionPlanRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Notification": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "userId": {
            "type": "integer",
            "title": "Userid"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "isRead": {
            "type": "boolean",
            "title": "Isread"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "userId",
          "title",
          "message",
          "type",
          "isRead",
          "createdAt"
        ],
        "title": "Notification"
      },
      "NotificationSettings": {
        "properties": {
          "emailAlerts": {
            "type": "boolean",
            "title": "Emailalerts"
          },
          "weeklyReports": {
            "type": "boolean",
            "title": "Weeklyreports"
          },
          "riskThreshold": {
            "type": "string",
            "title": "Riskthreshold"
          }
        },
        "type": "object",
        "required": [
          "emailAlerts",
          "weeklyReports",
          "riskThreshold"
        ],
        "title": "NotificationSettings"
      },
      "NotificationSettingsResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "settings": {
            "$ref": "#/components/schemas/NotificationSettings"
          }
        },
        "type": "object",
        "required": [
          "success",
          "settings"
        ],
        "title": "NotificationSettingsResponse"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "SuccessResponse"
      },
...
      "UserResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/app__schemas__UserRead"
          }
        },
        "type": "object",
        "required": [
          "user"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__schemas__UserRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "businessName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Businessname"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive",
            "default": true
          },
          "isSuperuser": {
            "type": "boolean",
            "title": "Issuperuser",
            "default": false
          },
          "isVerified": {
            "type": "boolean",
            "title": "Isverified",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "createdAt"
        ],
        "title": "UserRead"
      },
      "app__schemas__UserUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "businessName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Businessname"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          }
        },
        "type": "object",
        "title": "UserUpdate"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/login"
          }
        }
      }
    }
  }
}